requirements.txt
requests~=2.27.1
PyYAML~=6.0
Cerberus~=1.3.4